move blockchain class to its own folder
